requirements.txt
# requirements.txt for Internship Applications Automation

python-dotenv>=1.0.0

